backend/requirements.txt
# Core dependencies
flask==2.3.3
pydantic==2.4.2
motor==3.3.1  # MongoDB async driver
google-cloud-storage==2.10.0
firebase-admin==6.2.0
python-dotenv==1.0.0
boto3==1.28.44  # for AWS S3 